Summary moved to it's own controller.
