+++ app/views/applications/process/summary/index.html.slim
@@ -11,7 +11,7 @@ header
   -if !@application.benefits
     =build_section_with_defaults 'Income', @income, application_incomes_path(@application)
 
-= form_for @application, url: application_summary_save_path(@application), method: :put do |f|
+= form_for @application, url: application_summary_path(@application), method: :post do |f|
   = f.submit 'Complete processing', :class => 'button primary large'
